+++ libcob/termio.c
@@ -55,8 +55,8 @@ static void
 display_numeric (cob_field *f, FILE *fp)
 {
 	int		i;
-	int		digits;
-	int		scale;
+	unsigned short	digits;
+	signed short	scale;
 	int		size;
 	cob_field_attr	attr;
 	cob_field	temp;
@@ -92,8 +92,8 @@ pretty_display_numeric (cob_field *f, FILE *fp)
 	unsigned char	*p;
 	unsigned char	*q;
 	int		i;
-	int		digits;
-	int		scale;
+	unsigned short	digits;
+	signed short	scale;
 	int		size;
 	cob_field_attr	attr;
 	cob_field	temp;
